refactor: make openapi.Generate the config-aware schema entry point
Rename the config-aware schema generator package to pkg/schema/openapi so
call sites read openapi.Generate(cfg, dir). Rename the existing renderer
file to openapi_spec.go to free the openapi name.
