+++ pkg/schema/openapi_spec.go
