+++ pkg/schema/openapi_spec_test.go
